Fix and support for Conrad_RS-200 protocol (#313)
There is a tiny bug hat prevents the usage for "Conrad RS-200" based devices. The RS-200 protocol does not repeat packets that often. One need already to detect the end of the first packet. The gap between two transmission are not handled right when RF packet is received first time.

a) 1st gap detection fixed
b) Furthermore the "Conrad RS-200" protocol is added.