Send the syncFactor before the data
It is quite possible I totally misunderstood how to properly use this library, but I suspect there was a mistake there: the sync bits were not sent on the first repetition of the data and were appended at the end of the packet. It probably still worked as long as repeat >1 and that there is not interference.